backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
neo4j==5.14.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Redis and Task Queue
redis==5.0.1
celery==5.3.4

# AI/ML
openai==1.3.7
anthropic==0.7.7
langchain==0.0.340
langchain-openai==0.0.2
langchain-anthropic==0.0.1

# Speech-to-Text
openai-whisper==20231117
assemblyai==0.20.0

# Embeddings and Vector Operations
sentence-transformers==2.2.2
numpy==1.26.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Requests
httpx==0.25.2
aiohttp==3.9.1

# File Processing
python-magic==0.4.27
pillow==10.1.0

# Calendar Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1

# Monitoring and Logging
sentry-sdk==1.38.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
